requirements.txt
tensorflow==2.20.0
pandas 
dvc
mlflow
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-pyYAML
scipy
Flask
Flask-Cors
gdown
kagglehub
jupyter 
ipywidgets
tqdm